pyproject.toml
[tool.poetry]
name = "cookie-stand-admin-backend"
version = "0.1.0"
description = ""
authors = ["noureddein <noureddein@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
